Move scrapy/contrib/downloadermiddleware to scrapy/downloadermiddlewares
