+++ scrapy/downloadermiddlewares/decompression.py
