+++ scrapy/downloadermiddlewares/defaultheaders.py
