+++ scrapy/downloadermiddlewares/httpauth.py
