+++ scrapy/downloadermiddlewares/httpcache.py
