+++ scrapy/downloadermiddlewares/httpcompression.py
