+++ scrapy/downloadermiddlewares/redirect.py
