+++ scrapy/downloadermiddlewares/stats.py
