+++ tests/py3-ignores.txt
@@ -91,6 +91,9 @@ scrapy/contrib/linkextractors/htmlparser.py
 scrapy/contrib/downloadermiddleware/retry.py
 scrapy/contrib/downloadermiddleware/httpproxy.py
 scrapy/contrib/downloadermiddleware/cookies.py
+scrapy/downloadermiddlewares/retry.py
+scrapy/downloadermiddlewares/httpproxy.py
+scrapy/downloadermiddlewares/cookies.py
 scrapy/contrib/statsmailer.py
 scrapy/contrib/memusage.py
 scrapy/commands/deploy.py
