+++ tests/test_downloadermiddleware_defaultheaders.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 import six
 
-from scrapy.contrib.downloadermiddleware.defaultheaders import DefaultHeadersMiddleware
+from scrapy.downloadermiddlewares.defaultheaders import DefaultHeadersMiddleware
 from scrapy.http import Request
 from scrapy.spider import Spider
 from scrapy.utils.test import get_crawler
