+++ tests/test_downloadermiddleware_httpcompression.py
@@ -5,7 +5,7 @@ from gzip import GzipFile
 
 from scrapy.spider import Spider
 from scrapy.http import Response, Request, HtmlResponse
-from scrapy.contrib.downloadermiddleware.httpcompression import HttpCompressionMiddleware
+from scrapy.downloadermiddlewares.httpcompression import HttpCompressionMiddleware
 from tests import tests_datadir
 from w3lib.encoding import resolve_encoding
