+++ tests/test_downloadermiddleware_httpproxy.py
@@ -2,7 +2,7 @@ import os
 import sys
 from twisted.trial.unittest import TestCase, SkipTest
 
-from scrapy.contrib.downloadermiddleware.httpproxy import HttpProxyMiddleware
+from scrapy.downloadermiddlewares.httpproxy import HttpProxyMiddleware
 from scrapy.exceptions import NotConfigured
 from scrapy.http import Response, Request
 from scrapy.spider import Spider
