+++ tests/test_downloadermiddleware_robotstxt.py
@@ -3,7 +3,7 @@ import re
 from twisted.internet import reactor
 from twisted.internet.defer import Deferred
 from twisted.trial import unittest
-from scrapy.contrib.downloadermiddleware.robotstxt import RobotsTxtMiddleware
+from scrapy.downloadermiddlewares.robotstxt import RobotsTxtMiddleware
 from scrapy.exceptions import IgnoreRequest, NotConfigured
 from scrapy.http import Request, Response
 from scrapy.settings import Settings
